# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready AI Agent API with DDD architecture and Eino workflow engine. Supports OpenAI-compatible tool calling, human-in-the-loop approval, JWT authentication, and streaming responses.",
        "title": "Go Agent API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/wyuneed/go-agent-api/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/wyuneed/go-agent-api/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/tools": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of tools the authenticated token is allowed to use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "List available tools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/tools/batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Executes multiple tool calls concurrently and returns all results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Execute tools in parallel",
                "parameters": [
                    {
                        "description": "Array of tool calls",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_wyuneed_go-agent-api_pkg_toolspec.ToolCall"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerBatchResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/tools/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Executes a single tool call and returns the result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "Execute a tool",
                "parameters": [
                    {
                        "description": "Tool call",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_pkg_toolspec.ToolCall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerBatchResultData": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerBatchResultResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerBatchResultData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerErr"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolData": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolFunction"
                },
                "type": {
                    "type": "string",
                    "example": "function"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolFunction": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Perform mathematical calculations"
                },
                "name": {
                    "type": "string",
                    "example": "calculator"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolResultResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolsData": {
            "type": "object",
            "properties": {
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolData"
                    }
                }
            }
        },
        "github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_internal_pkg_response.SwaggerToolsData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_wyuneed_go-agent-api_pkg_toolspec.FunctionCall": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_wyuneed_go-agent-api_pkg_toolspec.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/github_com_wyuneed_go-agent-api_pkg_toolspec.FunctionCall"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer\" followed by a space and your JWT token. Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}